+++ BusinessServices/Interfaces/IFileBS.cs
@@ -3,9 +3,9 @@ using DriveAPI.Models;
 namespace DriveAPI.BussinesServices
 {
     public interface IFileBS{
-        public Filew GetFile(int id);
+        public Filew GetFile(long id);
         public Filew PostFile(Filew file);
-        public Filew PutFile(int id, Filew file);
-        public int DeleteFile(int id);
+        public Filew PutFile(long id, Filew file);
+        public long DeleteFile(long id);
     }    
 }
